When signing, preserve signatures that are still valid
When signing all the client files in preparation for a release, only
regenerate signatures that are missing or no longer valid, so we don't
have to update files on the clients that haven't actually changed.

+++ server/sign.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 
+import argparse
 import os
 import shutil
 import subprocess
@@ -10,19 +11,62 @@ from qlmdm import (
     set_gpg,
     release_files_iter,
     signatures_dir,
+    verify_signature,
     sign_file,
+    get_server_settings,
+    get_logger,
 )
 
 os.chdir(top_dir)
-set_gpg('server')
 
-shutil.rmtree(signatures_dir, ignore_errors=True)
+log = get_logger(get_server_settings(), 'sign')
 
-for file in release_files_iter():
-    sign_file(file)
 
-try:
-    subprocess.check_output(('python', os.path.join('client', 'verify.py')),
-                            stderr=subprocess.STDOUT)
-except:
-    sys.exit('Verify failed, try running {} again'.format(sys.argv[0]))
+def parse_args():
+    parser = argparse.ArgumentParser(description='Generate digital signatures '
+                                     'for client files')
+    parser.add_argument('--full', action='store_true', help='Regenerate all '
+                        'signatures rather than only invalid ones')
+
+    args = parser.parse_args()
+    return args
+
+
+def main():
+    args = parse_args()
+
+    old_signatures = set()
+    if args.full:
+        log.info('Renerating all signatures')
+        shutil.rmtree(signatures_dir, ignore_errors=True)
+    elif os.path.exists(signatures_dir):
+        strip = len(signatures_dir) + 1
+        for dirpath, dirnames, filenames in os.walk(signatures_dir):
+            for f in filenames:
+                old_signatures.add(os.path.join(dirpath, f))
+
+    for file in release_files_iter():
+        set_gpg('client')
+        signature = verify_signature(file)
+        if signature:
+            log.debug('Preserving valid signature for {}', file)
+        else:
+            set_gpg('server')
+            log.info('Signing {}', file)
+            signature = sign_file(file, overwrite=True)
+        old_signatures.discard(signature)
+
+    for file in old_signatures:
+        log.info('Removing obsolete signature {}', file)
+        os.unlink(file)
+
+    try:
+        subprocess.check_output(
+            ('python', os.path.join('client', 'verify.py')),
+            stderr=subprocess.STDOUT)
+    except:
+        sys.exit('Verify failed, try running {} again'.format(sys.argv[0]))
+
+
+if __name__ == '__main__':
+    main()
